Extract index pairs to separate class
With upcoming alternative voting formats, a voting round
should have only one at a time. But items, seed, and votes
will remain the same. Thus, only index pairs and functions
specific to it are extracted into the implementation-specific
VotingRound::ScoreBased structure. Index pairs are no longer
obtainable outside of VotingRound.

To facilitate this change, tests which relied on obtaining
index pairs instead either make use of the votes - which
will have the same information, but require the voting round
to be completed to be of use - or are removed as they were
testing implementation-specific code.

Since vote pruning is specific to the score-based voting format,
and since pruning should not occur during a voting round anyway,
it now only occurs during voting round creation. Tests which
verify this feature after creation are removed.

Also:
- Some test helper vector operators are moved to testing.h