(WIP) Add rank-based voting format
As an alternative to the existing score-based voting format,
which can be either full or reduced/partial, this adds
a rank-based voting format.

Full score-based voting scales the number of votes
with O(N^2), and does not guarantee a result without ties.
Reduced score-based voting scales the number of votes
with O(N), but also guarantees multiple ties.

Rank-based voting scales the number of votes with O(N*log(N)),
and guarantees no ties. It can however not be combined with others yet.